requirements.txt
requests
pyquery
beautifulsoup4
pyyaml
